pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "opendoku"
version = "1.0.0"
description = "A clean, open-source Sudoku desktop app built with Python and pywebview."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.13"
authors = [
  { name = "mietek64", url = "https://github.com/mietek64/opendoku" }
]
keywords = ["sudoku", "game", "desktop", "pywebview"]
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.13",
  "Operating System :: OS Independent",
  "Environment :: Win32 (MS Windows)",
  "Environment :: MacOS X",
  "Environment :: X11 Applications",
  "Topic :: Games/Entertainment :: Puzzle Games",
]
dependencies = [
  "pywebview>=4.4.0",
]

[project.optional-dependencies]
build = [
  "pyinstaller>=6.0.0",
]

[project.urls]
Homepage   = "https://github.com/mietek64/opendoku"
Repository = "https://github.com/mietek64/opendoku"
Issues     = "https://github.com/mietek64/opendoku/issues"
Changelog  = "https://github.com/mietek64/opendoku/blob/main/CHANGELOG.md"

[project.scripts]
opendoku = "main:main"

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
